+++ changelog.md
@@ -21,6 +21,7 @@
  * ForeignKey
  * struct Default(alias value_)
  * a lot of functions in utils.meta...
+ * SetConstraint
 
 ### Misc.